requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
PyPDF2==3.0.1
chromadb==0.6.3
langchain==0.3.19
langchain-openai==0.3.6
langchain-community==0.3.18
python-dotenv==1.0.1
openai==1.63.2